extracted Statistics component to a separate file
